Create renamesuffix.py
This small (and simple) code helps people rename the files downloded into a cumbersome suffix of .VIDEO_AUDIO into .avi for easy access. Will be a super useful tool to add to the repo as I felt is was needed. 
For the totally no-code approach to this repo :)

Doing some really good work here Nasir